TransactionSequence: cancel all subjobs on rollback-from-user.
Summary:
This replaces commit f14a6e997629 which wasn't the right fix for the
same problem.

Test Plan:
F5 in a large folder in kmail, then pressing the Abort button
in the middle of the sync. It used to just block the sync, while now it's
possible to resume afterwards.

Reviewers: dvratil, vkrause

Reviewed By: dvratil

Subscribers: kde-pim

Differential Revision: https://phabricator.kde.org/D19488