Also consider member typedefs for determining the base class
Summary:
This should allow us to avoid having to include (or copy) this into
KCalCore and KMime, as we no longer need to specialize the SuperClass
template.

PS: The unit test was written by Till 13 years ago, I have no idea who
Gerd and Rudi are, or why they are sub-classes of me.

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: dvratil, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D20845