Raise limit for PimItem.remoteId to 1024 characters
Summary:
Some DAV servers produce exteremely long RIDs that do not fit into the
current 255 character limit. This is only problem on MySQL, on other
DBs we use TEXT, which has no length limit.

On MySQL, just increasing the column length to 1024 is not enough,
there's a limit on maximum length of VARBINARY/VARCHAR index, which is
756 bytes (255 characters). To be able to have an index on a VARBINARY
column with up to 1024 characters we need to enable innodb_large_prefix
and change how the extra characters are stored on the disk (ROW_FORMAT).

BUG: 394839

Test Plan: Modified a unit-test to attempt creating an Item with 1024 characters long RID

Reviewers: #kde_pim, vkrause

Reviewed By: #kde_pim, vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D21714