Testrunner: wait until all jobs are done before aborting setup
Summary:

When resource setup fails for any reason, we must wait for the remaining
jobs to finish before shutting down Akonadi and terminating the testrunner.
Otherwise the next scheduled job gets executed and causes a new Akonadi
instance to be launched through DBus activation. This instance is not
managed by the testrunner though, so it does not ever get terminated,
causing ctest to hang.

Reviewers: dfaure

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D25458