Workaround broken Tag parentID in notifications
Invalid Tag has ID -1, but we store invalid parent in the DB as NULL
which gets auto-translated to 0 by our generated Entity code. To
work around that limitation we explicitly convert parent 0 to -1
when serializing Tag into response.

In the long run we need some proper fix, likely on the level of the
generated Entities.