Save configuration when creating resources for new user
Summary:
Akonadi calls `writeConfig` DBus method to save the configuration for newly
created resources, both via `firstrun` and `SpecialCollections` mechanisms.
This method is non-existent for all of the resources (it was deprecated in
KConfig, and apparently is not exported to DBus), method `save` should
be used instead.

This is related to issues raised in {D27905}: settings provided in firstrun did
not override default settings for the resource.

CCBUG: 345211
CCBUG: 352502

Test Plan:
1) Check `qdbus org.freedesktop.Akonadi.Resource.akonadi_maildir_resource_0 /Settings`
2) There is no `writeConfig` method, but there is `save` method

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D28523