ResourceBase: use erase-remove for getting list of valid items
In majority of cases all the items will be valid, so creating a
copy of the list by detaching is more efficient than copying it
item by item in a for loop. And for the cases when we need to
remove some items, the erase-remove idiom is the best choice.