AkonadiControl: expose agent instances of unknown types too
When an Agent type is removed/uninstalled so that after Akonadi restart
no information are available about the type, but agentsrc still contains
references to instances of that type there was no way for the user to
find or fix the problem as AkonadiControl would silently ignore the
instances.

This change makes AkonadiControl to reconstruct a dummy AgentType and
AgentInstance and expose it via DBus as usual so that they at least
in some form show up in the UI and the user can remove the account
or it can be picked up by the automated migration agent.