Server: fix fetching of attributes with empty data
There's a special path in ItemFetchHelper that skips an Item if the requested
part has an empty body and IgnoreError flag is set to true. This can of course
happen for payload parts, which can be expired, but is also perfectly valid
situation for attributes, which can simply have empty body by default, since
they never expire.

The whole logic there needs some revisiting, but that's for another patch and
task, since it has potential to break code that relies on all the
weirdness in the retrieval code.