Replace ResourceSelectJob with session cookies
When a new connection is established, the server generates a session cookie
that it sends back to the client. The client then sends the cookie with each
request to the server. The server associates certain information with a cookie,
for example when a Session is owned by an Akonadi Resource. Instead of using
the ResourceSelectJob, the information is instead sent by the Session during
the initial Login command.