Set package prefix dependent variables before calling find_dependency()
find_dependency() changes the package prefix variable to whatever that is
for the package we look for. That's commonly the same, so this doesn't
matter. In the Flatpak build environment however Akonadi and its
dependencies are in different prefixes, causing this to break modules
depending on Akonadi when they need any of these variables.