Implement CompressionStream
It's a QIODevice that can be placed on top of another
QIODevice. When opened in read-only mode, it will read
compressed data from the underlying device and output
decompressed data. When opened in write-only mode it
will take the data written into it and write them
compressed into the underlying device.