Introduce Config class, disable payload comp. by default
The Config is a global class that allows for run-time tuning
of behaviour of Akonadi clients, like enabling or disabling
payload compression, in future selecting compression algorithm
and compression level and possibly more.

At the same time this disables the payload compression feature
by default (can be enabled via config) as LZMA is too slow and
I failed to implement LZMA in time for 20.12.