Sort the "Recent Folder" list by most recently used
So that the most recently used folder appears at the top, and the
others follow in order.  Makes the list more usable for repeated
copies or moves to the same folder in KMail.  Also optimises the
case where the most recently used item is already at the top (no
need to write out config again).