Make the recently accessed folder list globally accessible
This is normally managed by RecentCollectionAction, but is only updated
by StandardActionManager when its standard actions are used.  If a
caller uses its own actions, for example KMail's separate message reader
window, then the list is not updated.

Add a static function StandardActionManager::addRecentCollection(),
which simply calls RecentCollectionAction::addRecentCollection() to
update the global recent collection list.  Read and update this list in
RecentCollectionAction as before.

CCBUG:435410