Remove the std::auto_ptr and std::unique_ptr payload checks
Now that we depend on C++17 std::auto_ptr doesn't exist anymore and thus
also cannot accidentally be put into an item payload.

The check for std::unique_ptr is unnecessary as it's not copyable (unlike
std::auto_ptr or a raw pointer), and as such will neither succeed to
compile with the item payload template code, nor will it be able to be
declared as a Qt metatype which is required for that.