Use all interface include dirs needed for LibXml2 (unbreak Windows build)
LIBXML2_INCLUDE_DIRS instead of LIBXML2_INCLUDE_DIR has all the needed
include dirs also of other libraries in the public interface.
iconv.h as used by libxml/encoding.h was not found on Windows due to that.

As CMake 3.16 is the min required version meanwhile, also resolve the
existing TODO and switch to the imported target LibXml2::LibXml2,
which has the DIRS variant set in the link interface property.