Move CalendarSupport::TagCache and tag color convenience methods here
This is in no way calendaring related, and not having it that high up in
the dependency chain will enable moving other things further down, which
should help with reducing widget-heavy dependencies on Kalendar for
example.

The code is substantially altered from CalendarSupport, combining TagCache
with (most of) the tag color API and the singleton behavior of KCalPrefs,
but functionality-wise it's still basically the same.