SQLite: don't force-enable shared cache mode
According to [SQLite documentation](https://www.sqlite.org/sharedcache.html)
the shared cache mode is deprecated and it's use is discouraged in favor
of the WAL mode.

I've been using Akonadi with this patch for a few days and I did not notice
any issues, in certain scenarios it actually feels a little bit faster, but
I have no numbers to back this claim.