Remove global data store transaction mutex
This is only used when running Akonadi against the upstream QSQLITE
driver and is actually not needed anymore. This was required back in the
early days of the Universe when SQLite did not support concurrent access
very well, so we used a global mutex to serialize all access. It was
only requied for the upstream QSQLITE driver, because the limitations
have been worked around in the Akonadi QSQLITE3 fork of the driver.

Nowadays SQLite supports concurrency without any problems and since we
use the WAL mode now, the upstream QSQLITE driver will work as well as
our fork.