Split parts of AgentConfigurationBase into AgentWidgetConfigurationBase
This is done in preparation for support for QML-based agent configuration
UI. For this we need to move widget-specific logic and dependencies into
AkonadiWidgets. While this is an API breaking change, it only affects a
couple of resources in kdepim-runtime and kmail that have been using the
"new" configuration plugin system.