Fix crash on server shutdown
When AkonadiServer::quit() is called it clears the vector of existing
connections and destroys them. This can race with a resource disconnecting
or crashing on its own, which then enqueues a singal emission from the
Connection to remove itself from the connections vector as well.

Previously the code assumed the Connection must always exist in the vector
which caused a crash when it did not...surprise surprise.

BUG: 462692
FIXED-IN: 23.04.0