QueryBuilder: make it possible to pass a custom DataStore
Add a constructor overload that takes DataStore pointer as
an argument and uses it to obtain the QSqlDatabase to run
queries against, instead of using the DataStore::self()
singleton.

This is the very first step towards getting rid of DataStore
as a per-thread singleton, which makes testing difficult,
and to also make it possible to use the existing infrastructure
to implement a migrator between two different database backends.