Entities: add overloads that accept DataStore
Add overloads that accept DataStore pointer argument to all
methods in Entity (and all entities generated from schema) that
interact with the database.

This allows controlling which DataStore entities should use to
interact with the DB and is the next step towards getting rid
of DataStore as a per-thread singleton.