Fix ItemCreateHandlerTest
The list of new Item's parts is stored in QSet, which has randomized order
by default. This breaks the test since the test scenarios are hard-coded
and expect a specific order in which the server requests the parts.

Since introducing variable scenarios would make the tests overcomplicated,
we just settle with setting a stable QHash seed to make the QSet ordering
deterministic in the test.