Monitor: restore Items from Ntf if we cannot fetch them
The logic previous was flawed as it was possible to bypass the !items.isEmpty()
check if the notification was a removal notification. This would later crash
in emitItemNotification() which would try to access the first element of the
(empty) item list.

We now always fall back to reconstructing the Item list from the notification
if we failed to retrieve anything.